+++ src/design/components/index.ts
@@ -1,0 +1,5 @@
+export * from './Container';
+export * from './Divider';
+export * from './IconButton';
+export * from './PageHeader';
+export * from './Section';